Deal with symlinks intelligently
Symlinks that are not "owned" by a torrent file are considered extra
data.

Using symlinks within the data directory to provide multiple ways of
accessing the same data (for example, the data is terribly named and you
want to access it in a sane manner) caused the scanner to consider the
symlink extra data since it's not owned by any torrent file.

This commit fixes this by adding a dedicated symlink section of the
report that filters out any symlinks that target data in the data
directory. Only symlinks not owned by any torrent and not pointing to a
file within the data directory will be flagged in the output.